SagTensionAnalysisResult file renamed to match class.
