+++ include/analysis_controller.h
@@ -10,7 +10,7 @@
 #include "models/sagtension/line_cable_unloader.h"
 #include "wx/wx.h"
 
-#include "sag_tension_analysis_results.h"
+#include "sag_tension_analysis_result.h"
 #include "span.h"
 
 /// \par OVERVIEW
